requirements.txt
streamlit>=1.28,<2.0
numpy>=1.24,<3.0
pillow>=10.0,<12.0
requests>=2.31,<3.0
